refactor: Organize frontend structure with pages folder and shared types
- Move all page components from src/ to src/pages/ folder
- Update relative imports in page components to use ../ paths
- Add shared ImageItem and StyleTransfer interfaces to types/ui.ts
- Update ImageFeed and ImageModal to import from shared types
- Remove duplicate interface definitions
