+++ frontend/src/pages/ImageEdit.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
-import { Label, Field, Section } from "./components/UI";
-import { apiClient } from "./lib/apiClient";
-import ImageFeed from "./components/ImageFeed";
-import { useSmartResolution } from "./hooks/useSmartResolution";
+import { Label, Field, Section } from "../components/UI";
+import { apiClient } from "../lib/apiClient";
+import ImageFeed from "../components/ImageFeed";
+import { useSmartResolution } from "../hooks/useSmartResolution";
 
 type Tab = "edit" | "camera-angle";
