+++ frontend/src/pages/MultiTalkOnePerson.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
-import { startJobMonitoring, checkComfyUIHealth } from "./components/utils";
-import VideoFeed from "./components/VideoFeed";
-import { useSmartResolution } from "./hooks/useSmartResolution";
-import { apiClient } from "./lib/apiClient";
+import { startJobMonitoring, checkComfyUIHealth } from "../components/utils";
+import VideoFeed from "../components/VideoFeed";
+import { useSmartResolution } from "../hooks/useSmartResolution";
+import { apiClient } from "../lib/apiClient";
 
 // VAPAI One-Person Frontend for ComfyUI
 // - Enter ComfyUI URL
